requirements.txt
Django==1.7.8
djangorestframework==3.1.1
MySQL-python==1.2.5
beautifulsoup4==4.5.1
lxml==3.6.4
requests==2.11.1
six==1.10.0
wsgiref==0.1.2
